mv cb test to ica
